requirements.txt
Flask
requests
python-dotenv
Flask-SQLAlchemy
psycopg2-binary
werkzeug
Flask-JWT-Extended
gunicorn